Refactor code before merge

+++ src/tanks/mediumtank.h
@@ -10,8 +10,8 @@ public:
     ~MediumTank();
 
     // TODO: Implement methods
-    bool CanShoot(const Vector3d& point) const override;
-    bool CanMove(const Vector3d& point) const override;
+    bool CanShoot(const Vector3i& point) const override;
+    bool CanMove(const Vector3i& point) const override;
 };
 
 #endif // SERVER_MEDIUMTANK_H
